requirements.txt
streamlit
requests
pillow
python-dotenv
pandas
numpy
oracledb
sentence-transformers
oci
cohere
langchain
langchain-community
langchain-core
langgraph
PyPDF2
openpyxl
xlrd
